+++ src/Bugfree.Spo.Analytics.Cli.Tests/AssemblyInfo.fs
@@ -1,10 +1,10 @@
-﻿namespace Bugfree.Spo.Analytics.Cli.AssemblyInfo
+﻿namespace Bugfree.Spo.Analytics.Cli.Tests.AssemblyInfo
 
 open System.Reflection
 
-[<assembly: AssemblyTitle("Bugfree.Spo.Analytics.Cli")>]
+[<assembly: AssemblyTitle("Bugfree.Spo.Analytics.Cli.Tests")>]
 [<assembly: AssemblyDescription("")>]
-[<assembly: AssemblyProduct("Bugfree.Spo.Analytics.Cli")>]
+[<assembly: AssemblyProduct("Bugfree.Spo.Analytics.Cli.Tests")>]
 [<assembly: AssemblyCopyright("Copyright (c) 2016-2017 Ronnie Holm <mail@bugfree.dk>")>]
 [<assembly: AssemblyVersion("1.2.0")>]
 [<assembly: AssemblyFileVersion("1.2.0")>]